removed setState() and structured form validation code
