+++ lib/widgets/layout/student_list_card.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 import 'package:flutter/material.dart';
-import '../models/student_model.dart';
-import '../widgets/confirm_dialog.dart';
+import '../../models/student_model.dart';
+import '../form/common/confirm_dialog.dart';
 
 class StudentListCard extends StatelessWidget {
   final StudentModel student;
