--- lib/widgets/student_form.dart
@@ -1,168 +1,0 @@
-import 'dart:io';
-import 'package:flutter/material.dart';
-import 'package:image_picker/image_picker.dart';
-import 'package:provider/provider.dart';
-import '../models/student_model.dart';
-import '../providers/student_provider.dart';
-
-class StudentForm extends StatefulWidget {
-  final StudentModel? student;
-
-  const StudentForm({super.key, this.student});
-
-  @override
-  State<StudentForm> createState() => _StudentFormState();
-}
-
-class _StudentFormState extends State<StudentForm> {
-  final _formKey = GlobalKey<FormState>();
-  late TextEditingController _nameController;
-  late TextEditingController _ageController;
-  late TextEditingController _emailController;
-  late TextEditingController _phoneController;
-  String? _imagePath;
-
-  @override
-  void initState() {
-    super.initState();
-    _nameController = TextEditingController(text: widget.student?.name ?? '');
-    _ageController = TextEditingController(
-      text: widget.student?.age.toString() ?? '',
-    );
-    _emailController = TextEditingController(text: widget.student?.email ?? '');
-    _phoneController = TextEditingController(
-      text: widget.student?.phone.toString() ?? '',
-    );
-    _imagePath = widget.student?.imagePath;
-  }
-
-  @override
-  void dispose() {
-    _nameController.dispose();
-    _ageController.dispose();
-    _emailController.dispose();
-    _phoneController.dispose();
-    super.dispose();
-  }
-
-  Future<void> _pickImage() async {
-    final picker = ImagePicker();
-    final pickedFile = await picker.pickImage(source: ImageSource.gallery);
-    if (pickedFile != null) {
-      setState(() => _imagePath = pickedFile.path);
-    }
-  }
-
-  void _saveStudent() {
-    if (!_formKey.currentState!.validate() || _imagePath == null) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(
-          content: Text('Please fill all fields and select an image'),
-        ),
-      );
-      return;
-    }
-
-    final student = StudentModel(
-      id: widget.student?.id,
-      name: _nameController.text.trim(),
-      age: int.parse(_ageController.text.trim()),
-      email: _emailController.text.trim(),
-      phone: int.parse(_phoneController.text.trim()),
-      imagePath: _imagePath!,
-    );
-
-    final provider = Provider.of<StudentProvider>(context, listen: false);
-    if (widget.student == null) {
-      provider.addStudent(student);
-    } else {
-      provider.updateStudent(student);
-    }
-
-    ScaffoldMessenger.of(context).showSnackBar(
-      const SnackBar(content: Text('Student saved successfully!')),
-    );
-    Navigator.pop(context);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final isEditing = widget.student != null;
-
-    return SingleChildScrollView(
-      child: Form(
-        key: _formKey,
-        child: Column(
-          children: [
-            GestureDetector(
-              onTap: _pickImage,
-              child: CircleAvatar(
-                radius: 50,
-                backgroundImage:
-                    _imagePath != null ? FileImage(File(_imagePath!)) : null,
-                child:
-                    _imagePath == null
-                        ? const Icon(Icons.add_a_photo, size: 30)
-                        : null,
-              ),
-            ),
-            const SizedBox(height: 16),
-            TextFormField(
-              controller: _nameController,
-              decoration: const InputDecoration(labelText: 'Name'),
-              validator:
-                  (value) =>
-                      value == null || value.isEmpty ? 'Enter name' : null,
-            ),
-            const SizedBox(height: 10),
-            TextFormField(
-              controller: _ageController,
-              decoration: const InputDecoration(labelText: 'Age'),
-              keyboardType: TextInputType.number,
-              validator: (value) {
-                final age = int.tryParse(value ?? '');
-                if (age == null || age <= 0 || age > 100) {
-                  return 'Enter valid age';
-                }
-                return null;
-              },
-            ),
-            const SizedBox(height: 10),
-            TextFormField(
-              controller: _emailController,
-              decoration: const InputDecoration(labelText: 'Email'),
-              keyboardType: TextInputType.emailAddress,
-              validator: (value) {
-                final emailRegex = RegExp(
-                  r"^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$",
-                );
-                if (value == null || value.isEmpty) {
-                  return 'Email is required';
-                } else if (!emailRegex.hasMatch(value)) {
-                  return 'Enter a valid email address';
-                }
-                return null;
-              },
-            ),
-            const SizedBox(height: 10),
-            TextFormField(
-              controller: _phoneController,
-              decoration: const InputDecoration(labelText: 'Phone'),
-              keyboardType: TextInputType.phone,
-              validator:
-                  (value) =>
-                      value != null && value.length == 10
-                          ? null
-                          : 'Phone must be 10 digits',
-            ),
-            const SizedBox(height: 20),
-            ElevatedButton(
-              onPressed: _saveStudent,
-              child: Text(isEditing ? 'Update' : 'Save'),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
